requirements.txt
finnhub-python==2.4.22
py-cord==2.6.1
python-dotenv==1.0.1
requests==2.22.0
urllib3==1.22.0
pandas==1.1.3
pyodbc<=4.0.0
pyzmq==21.0.0
matplotlib==3.5.2
numpy<=1.24.4
pynacl<=1.5.0
discord.py<=2.4.0
